refactor(drive): split service implementation into multiple files
