+++ modules/drive/app/index.go
@@ -1,10 +1,14 @@
 package app
 
-import "github.com/sky-as-code/nikki-erp/common/deps_inject"
+import (
+	"github.com/sky-as-code/nikki-erp/common/deps_inject"
+	drive_file "github.com/sky-as-code/nikki-erp/modules/drive/app/drive_file_service_impl"
+	drive_file_share "github.com/sky-as-code/nikki-erp/modules/drive/app/drive_file_share_service_impl"
+)
 
 func InitServices() error {
 	return deps_inject.Register(
-		NewDriveFileService,
-		NewDriveFileShareService,
+		drive_file.NewDriveFileService,
+		drive_file_share.NewDriveFileShareService,
 	)
 }
